apps/api/requirements.txt
# FastAPI Backend - Job Intelligence Platform

## Requirements
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.3.2

# Redis / Caching
redis==5.0.7
slowapi==0.1.9

# NLP & Embeddings
spacy==3.7.5
sentence-transformers==3.0.1
rapidfuzz==3.9.4

# Resume Parsing
PyMuPDF==1.24.7
python-docx==1.1.2
dateparser==1.2.0

# AI Providers (multi-provider support)
anthropic==0.31.0
google-generativeai==0.7.2
openai==1.35.7

# Celery
celery[redis]==5.4.0

# Encryption
cryptography==42.0.8

# HTTP
httpx==0.27.0
feedparser==6.0.11

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
